requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.10.1
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.7.0
